remove runner-related features from the main perfer process and move them to the runner
